Move read_metadata function to service-specific utilities
